refactor: refactored nova specific functions to a separate folder
